+++ src/shuriken/namespace.clj
@@ -1,8 +1,9 @@
-(ns shuriken.namespace)
+(ns shuriken.namespace
+  (:require [clojure.string :as str]))
 
 (defn- class-name [klass]
   (-> (str klass)
-      (clojure.string/replace #"(?:class|interface) " "")
+      (str/replace #"(?:class|interface) " "")
       symbol))
 
 (defn ns-syms [ns]
@@ -19,28 +20,45 @@
   ;; ns   . or /  name  eol
   #"(^.*)(?:/|\.)([^./]+)$")
 
-(defn fully-qualify
-  "Returns a fully-qualified symbol as if resolved from within a namespace.
-  Handles namespace aliases. Default namespace is *ns*.
+(defn- could-be-class-name? [x]
+  (let [x (str x)]
+    (and (-> x
+             (str/split #"\.")
+             last
+             (.codePointAt 0)
+             Character/isUpperCase)
+         x)))
 
-  (fully-qualified? 'IRecord)       => clojure.lang.IRecord
-  (fully-qualified? 'my-var)        => my-ns/my-var
-  (fully-qualified? 'alias/my-var)  => actual.namespace/my-var"
+(defn fully-qualify
+  "Returns the fully-qualified form of the symbol as if resolved from within a
+  namespace.
+  Handles namespace aliases.
+  Default namespace is *ns*.
+  
+  (fully-qualified? 'IRecord)      => clojure.lang.IRecord
+  (fully-qualified? 'my-var)       => my-ns/my-var
+  (fully-qualified? 'alias/my-var) => actual.namespace/my-var"
   ([sym]
    (fully-qualify *ns* sym))
   ([ns sym]
-    (if-let [r (ns-resolve *ns* sym)]
-      (if (class? r)
-        (class-name r)
-        (-> r meta :ns str
-            (str "/" (.sym r))
-            symbol))
-      sym)))
+   (if (some-> (.getNamespace sym) could-be-class-name?)
+     (symbol (str (-> (.getNamespace sym) symbol fully-qualify)
+                  \/
+                  (-> sym .getName symbol)))
+     (if-let [r (ns-resolve *ns* sym)]
+       (if (class? r)
+         (class-name r)
+         (symbol (str (-> r meta :ns str)
+                      \/
+                      (.sym r))))
+       sym))))
 
 (defn fully-qualified?
   "Returns true if the symbol constitutes an absolute reference.
+  See fully-qualified.
+  
   Handles namespace aliases.
-  ns defaults to *ns*.
+  Default namespace is *ns*.
 
   (fully-qualified? 'clojure.lang.IRecord) => true
   (fully-qualified? 'my-ns/my-var)         => true
@@ -54,27 +72,29 @@
           (re-matches fully-qualified-regex (str sym))))))
 
 (defn unqualify
-  "Returns the unqualified version of sym.
+  "Returns the unqualified form of sym.
   Handles namespace aliases.
   ns defaults to *ns*.
 
-  (unqualifiy 'clojure.lang.IRecord) => IRecord
-  (unqualifiy 'my-ns/my-var)         => my-var
-  (unqualifiy 'alias/my-var)         => alias/my-var"
+  (unqualifiy 'clojure.lang.IRecord)      => IRecord
+  (unqualifiy 'my-ns/my-var)              => my-var
+  (unqualifiy 'alias/my-var)              => alias/my-var
+  (unqualifiy 'some.path.Class/staticMeth => Class/staticMeth"
   ([sym]
    (unqualify *ns* sym))
   ([ns sym]
    (if (fully-qualified? sym)
      (let [[_ ns-part name] (re-matches fully-qualified-regex (str sym))
-           shortcut (-> (->> ns ns-aliases
-                             (map (fn [[sym ns]]
-                                    [(-> ns str symbol)
-                                     sym]))
-                             (into {}))
-                        (get (symbol ns-part)))]
+           shortcut (or (some-> ns-part could-be-class-name?)
+                        (-> (->> ns ns-aliases
+                                 (map (fn [[sym ns]]
+                                        [(-> ns str symbol)
+                                         sym]))
+                                 (into {}))
+                            (get (symbol ns-part))))]
        (if (and shortcut
                 (not (contains? (ns-syms ns) sym)))
-         (symbol (str shortcut \/ name))
+         (symbol (str (unqualify (symbol shortcut)) \/ name))
          (symbol name)))
      sym)))